adk-server/pyproject.toml
[project]
name = "marketlogic-adk"
version = "0.1.0"
description = "ADK service for MarketLogic"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.111",
  "uvicorn[standard]>=0.30",
  "pydantic>=2.7",
  "pydantic-settings>=2.3",
  "python-dotenv>=1.0",
  "google-adk>=1.0",
  "asyncpg>=0.30",
  "sqlalchemy>=2.0",
  "loguru>=0.7",
  "httpx>=0.27",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-asyncio>=0.23",
  "httpx>=0.27",
  "ruff>=0.5",
  "mypy>=1.10",
]

[tool.uv]
managed = true

[tool.ruff]
line-length = 100

[tool.mypy]
python_version = "3.12"
strict = true
